Working docker based state machine

+++ drove-models/src/main/java/com/phonepe/drove/models/application/requirements/MemoryRequirement.java
@@ -11,7 +11,7 @@ import lombok.Value;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 public class MemoryRequirement extends ResourceRequirement {
-    int sizeInMB;
+    long sizeInMB;
 
     public MemoryRequirement(int sizeInMB) {
         super(ResourceRequirementType.MEMORY);
